// src/store/action-types/user-profile/user-profile-action-type.ts
export enum UserProfileActionType {
    AVATAR_CHANGED = "AVATAR_CHANGED",
    USER_METADATA_LOADED = "USER_METADATA_LOADED",
    USER_PROFILE_LOADING = "USER_PROFILE_LOADING",
    USER_PROFILE_NOT_LOADING = "USER_PROFILE_NOT_LOADING",
    USER_AVATAR_CHANGE_LOADING = "USER_AVATAR_CHANGE_LOADING",
    RESET_USER_PROFILE_STATE = "RESET_USER_PROFILE_STATE",
    AVATAR_CHANGE_ERRORED = "AVATAR_CHANGE_ERRORED",
    AVATAR_CHANGE_ERROR_HANDLED = "AVATAR_CHANGE_ERROR_HANDLED"
}